my_flask_app/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.3
sib-api-v3-sdk==7.6.0
gunicorn==23.0.0
razorpay==1.4.2
